16:54/09/12/2017

+++ hc_sr04/src/main.rs
@@ -38,7 +38,7 @@ fn calc_distance(output: u64, input: u64) -> sysfs_gpio::Result<()> {
 
         let time = start.elapsed().subsec_nanos();
 
-        println!("Distance: {}mm", time * 0.171500);
+        println!("Distance: {}mm", time as f64 * 0.171500);
 
         Ok(())
     })
